Enable Bazel 8.1.1 to compile with musl
Bazel is built to compile with glibc. That means it does not compile
out of the box with Alpine Linux, which is intended to be used with musl
due to its smaller binary size.

However, it's very close. This change makes Bazel 8.1.1 compile with
musl:

* replace off64_t with off_t
* replace stat64 with stat
* replace fstatat64 with fstatat

A bit of research says that all of this is fine for glibc-based
platforms as well as long as you set _FILE_OFFSET_BITS=64. But I don't
care about those for now.